Graphics.SvgTree

Description

Module providing basic input/output for the SVG document, for document building, please refer to Graphics.Svg.Types.

Saving/Loading functions

loadSvgFile :: FilePath -> IO (Maybe Document)

Try to load an svg file on disc and parse it as a SVG Document.

parseSvgFile

Arguments

:: FilePath
Source path/URL of the document, used to resolve relative links.

-> Text

-> Maybe Document

Parse an in-memory SVG file

parseSvg :: Text -> Tree

unparse :: Element -> Tree

xmlOfDocument :: Document -> Element

Transform a SVG document to a XML node.

xmlOfTree :: Tree -> Maybe Element

saveXmlFile :: FilePath -> Document -> IO ()

Save a svg Document to a file on disk.

Manipulation functions

cssApply :: [CssRule] -> Tree -> Tree

Rewrite a SVG Tree using some CSS rules.

This action will propagate the definition of the css directly in each matched element.

cssRulesOfText :: Text -> [CssRule]

Parse CSS text into rules.

Type definitions

module Graphics.SvgTree.Types