Graphics.SvgTree.CssTypes

Description

Defines the types used to describes CSS elements

Documentation

data CssSelector

Defines complex selector.

Constructors

Nearby
Corresponds to the + CSS selector.

DirectChildren
Corresponds to the > CSS selectro.

AllOf [CssDescriptor]
Grouping construct, all the elements of the list must be matched.

type CssSelectorRule = [CssSelector]

A CssSelectorRule is a list of all the elements that must be met in a depth first search fashion.

data CssRule

Represents a CSS selector and the different declarations to apply to the matched elemens.

Constructors

CssRule

Fields

cssRuleSelector :: ![CssSelectorRule]

At the first level represents a list of elements to be matched. If any match is made, you can apply the declarations. At the second level

cssDeclarations :: ![CssDeclaration]

Declarations to apply to the matched element.

data CssDescriptor

Describes an element of a CSS selector. Multiple elements can be combined in a CssSelector type.

Constructors

OfClass Text
.IDENT

OfName Text
IDENT

OfId Text
#IDENT

OfPseudoClass Text
`:IDENT` (ignore function syntax)

AnyElem
*

WithAttrib Text Text
``

data CssDeclaration

Represents the content to apply to some CSS matched rules.

Constructors

CssDeclaration

Fields

_cssDeclarationProperty :: Text

Property name to change (like font-family or color).

_cssDecarationlValues :: [[CssElement]]

List of values

data CssElement

Value of a CSS property.

Constructors

CssIdent !Text

CssString !Text

CssReference !Text

CssNumber !Number

CssColor !PixelRGBA8

CssFunction !Text ![CssElement]

CssOpComa

CssOpSlash

class CssMatcheable a where

Interface for elements to be matched against some CssRule.

Methods

cssIdOf :: a -> Maybe Text

For an element, tell its optional ID attribute.

cssClassOf :: a -> [Text]

For an element, return all of it's class attributes.

cssNameOf :: a -> Text

Return the name of the tagname of the element

cssAttribOf :: a -> Text -> Maybe Text

Return a value of a given attribute if present

type CssContext a = [[a]]

Represents a zipper in depth at the first list level, and the previous nodes at in the second list level.

type Dpi = Int

Alias describing a "dot per inch" information used for size calculation (see toUserUnit).

data Number

Encode complex number possibly depending on the current render size.

Constructors

Num Double
Simple coordinate in current user coordinate.

Px Double
With suffix "px"

Em Double
Number relative to the current font size.

Percent Double
Number relative to the current viewport size.

Pc Double

Mm Double
Number in millimeters, relative to DPI.

Cm Double
Number in centimeters, relative to DPI.

Point Double
Number in points, relative to DPI.

Inches Double
Number in inches, relative to DPI.

serializeNumber :: Number -> String

Encode the number to string which can be used in a CSS or a svg attributes.

findMatchingDeclarations :: CssMatcheable a => [CssRule] -> CssContext a -> [CssDeclaration]

Given CSS rules, find all the declarations to apply to the element in a given context.

toUserUnit :: Dpi -> Number -> Number

This function replaces all device dependant units with user units given its DPI configuration. Preserves percentage and "em" notation.

mapNumber :: (Double -> Double) -> Number -> Number

Helper function to modify inner value of a number

tserialize :: TextBuildable a => a -> Builder

Serialize an element to a text builder.